third_party/WebKit/LayoutTests/fast/html/imports/import-shared-remove.html

<!DOCTYPE html>
<html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<head>
<link id="main" rel="import" href="resources/hello.html">
<link id="shared1" rel="import" href="resources/hello.html">
<link id="shared2" rel="import" href="resources/hello.html">
</head>
<body>
<script>
test(function() {
    var main = document.getElementById("main");
    var shared1 = document.getElementById("shared1");
    var shared2 = document.getElementById("shared2");

    assert_true(main.import != null);
    assert_true(shared1.import != null);
    assert_true(shared2.import != null);

    document.head.removeChild(shared1);
    assert_true(main.import != null);
    assert_true(shared1.import == null);
    assert_true(shared2.import != null);

    document.head.removeChild(main);
    assert_true(main.import == null);
    assert_true(shared1.import == null);
    assert_true(shared2.import != null);

},
"Check that change on a link with same url doesn't affect others."
);
</script>
</body>
</html>
